Move NVT specifics to nvt crate within sink
