+++ rust/nasl-syntax/src/lib.rs
@@ -20,7 +20,7 @@ pub use token::Token;
 pub use token::StringCategory;
 pub use token::Base as NumberBase;
 pub use token::Keyword;
-pub use sink::ACT as ACT;
+pub use sink::nvt::ACT as ACT;
 
 
 /// Parses given code and returns found Statements and Errors
